implements validation/expiry of token

+++ src/main/java/com/senacor/controller/SpeechController.java
@@ -2,6 +2,7 @@ package com.senacor.controller;
 
 
 import com.senacor.model.Speech;
+import com.senacor.service.AuthenticationService;
 import com.senacor.service.SpeechService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.Link;
@@ -22,53 +23,70 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 public class SpeechController {
 
     private final SpeechService speechService;
+    private final AuthenticationService authenticationService;
 
     @Autowired
-    public SpeechController(SpeechService speechService) {
+    public SpeechController(SpeechService speechService, AuthenticationService authenticationService) {
         this.speechService = speechService;
+        this.authenticationService = authenticationService;
     }
 
     @RequestMapping(value = "", method = RequestMethod.GET)
-    public List<Speech> getSpeechesForEvent(@PathVariable(value = "eventId") String eventID) {
-        return speechService.getAllSpeechesForEvent(eventID);
+    public List<Speech> getSpeechesForEvent( @PathVariable(value = "eventId") String eventID) {
+            return speechService.getAllSpeechesForEvent(eventID);
     }
 
     @RequestMapping(value = "/{speechID}", method = RequestMethod.GET)
-    public Speech getSpeech(@PathVariable(value = "eventId") String eventID, @PathVariable("speechID")String speechID){
-        return speechService.getSpeech(eventID, speechID);
+    public ResponseEntity<Speech> getSpeech(@RequestHeader("Authorization") String tokenId, @PathVariable(value = "eventId") String eventID, @PathVariable("speechID")String speechID) {
+        if (authenticationService.isAuthenticatedUser(tokenId)) {
+            return new ResponseEntity<Speech>(speechService.getSpeech(eventID, speechID), HttpStatus.OK);
+        }else{
+            return new ResponseEntity<Speech>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(value = "/{speechID}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteSpeech(@PathVariable(value = "eventId") String eventID,@PathVariable("speechID")String speechID ) {
-        speechService.deleteSpeech(eventID, speechID);
+    public void deleteSpeech(@RequestHeader("Authorization") String tokenId, @PathVariable(value = "eventId") String eventID,@PathVariable("speechID")String speechID ) {
+            if (authenticationService.isAuthenticatedUser(tokenId)) {
+
+                speechService.deleteSpeech(eventID, speechID);
+            }
     }
 
     @RequestMapping(value = "/createSpeech", method = RequestMethod.POST)
-    public ResponseEntity addSpeech(@PathVariable(value = "eventId") String eventID, @RequestBody Speech speech) {
-        System.out.println("speechcontroller: addspeech method reached");
-        Speech savedSpeech = speechService.addSpeech(eventID, speech);
-        if (savedSpeech != null) {
-            return new ResponseEntity(savedSpeech, HttpStatus.CREATED);
+    public ResponseEntity addSpeech(@RequestHeader("Authorization") String tokenId, @PathVariable(value = "eventId") String eventID, @RequestBody Speech speech) {
+        if (authenticationService.isAuthenticatedUser(tokenId)) {
+
+            Speech savedSpeech = speechService.addSpeech(eventID, speech);
+            if (savedSpeech != null) {
+                return new ResponseEntity(savedSpeech, HttpStatus.CREATED);
+            } else {
+                HttpHeaders headers = new HttpHeaders();
+                headers.add("409-Status-Reason: ", "Validation failed");
+                return new ResponseEntity(speech, headers, HttpStatus.CONFLICT);
+            }
         }else{
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("409-Status-Reason: ",  "Validation failed");
-            return new ResponseEntity(speech, headers, HttpStatus.CONFLICT);
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
     }
 
     @RequestMapping(value = "/{speechId}", method = RequestMethod.PUT)
-    public ResponseEntity<Speech> editSpeech(@PathVariable(value = "eventId") String eventID, @RequestBody Speech speech) {
-        System.out.println("speechcontroller: in editspeech method");
-        Speech editedSpeech = speechService.editSpeech(eventID, speech);
-        if (editedSpeech != null) {
-            return new ResponseEntity<>(editedSpeech, HttpStatus.OK);
+    public ResponseEntity<Speech> editSpeech(@RequestHeader("Authorization") String tokenId, @PathVariable(value = "eventId") String eventID, @RequestBody Speech speech) {
+        if (authenticationService.isAuthenticatedUser(tokenId)) {
+
+            Speech editedSpeech = speechService.editSpeech(eventID, speech);
+            if (editedSpeech != null) {
+                return new ResponseEntity<>(editedSpeech, HttpStatus.OK);
+            } else {
+                HttpHeaders headers = new HttpHeaders();
+                headers.add("409-Status-Reason: ", "Validation failed");
+                Link selflink = linkTo(EventController.class).slash(eventID + "/speeches/" + speech.getSpeechId()).withSelfRel();
+                speech.add(selflink);
+                return new ResponseEntity<>(speech, headers, HttpStatus.CONFLICT);
+            }
         }else{
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("409-Status-Reason: ", "Validation failed");
-            Link selflink = linkTo(EventController.class).slash(eventID + "/speeches/" + speech.getSpeechId()).withSelfRel();
-            speech.add(selflink);
-            return new ResponseEntity<>(speech, headers, HttpStatus.CONFLICT);
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
     }
